+++ src/StartupOptionsWindow.h
@@ -1,7 +1,0 @@
-#include <proto/dos.h>
-
-extern Object *startupOptionsWindowObject;
-
-LONG createStartupOptionsWindow();
-
-void addStartupOptionsWindowActions();